fix(governance): expose action_scope in plan/trace output, close two gaps (#111)
Post-merge review of the action_scope.spend model (#107) and skill_eligibility
gate (#100/#101) found three related issues:

- action_scope (including spend.max_spend/allowed_vendors) was parsed onto
  Unit but never survived into PlannedUnit or UnitTrace — the shapes actually
  serialized as plan/trace JSON. A downstream enforcer (or a human running
  `kcp-agent trace --json`) had no way to see a selected skill's declared
  action scope without re-fetching and re-parsing the raw manifest
  independently, which both duplicates parsing logic and can drift from what
  was actually evaluated if the manifest changes between plan and invoke.
  kcp-harness's own governor.ts carries a comment describing exactly this
  workaround. Added action_scope to both PlannedUnit and UnitTrace (selected
  units only). Verified against every existing plan/trace conformance fixture
  (Java's PlanJsonConformanceTest/TraceJsonConformanceTest/vectors corpus) —
  none currently exercise a selected skill's action_scope, so this is a pure
  addition with no byte-for-byte drift. Java/Rust's own plan/trace JSON
  serializers don't emit it yet — noted as follow-up, not silently assumed.

- Only an exact `kind === "skill"` string matched the governed-unit path,
  with no validation anywhere on the kind field. A typo or unsupported value
  (kind: Skill, kind: procedure) silently loaded as an ordinary, ungoverned
  knowledge unit — skill_eligibility never applied. Added a validate.ts
  warning (mirroring the existing unknown-access-value check).

- The Java port's ManifestParser.asNum returned Long, truncating
  action_scope.spend.max_spend's fraction (4.99 -> 4) — a cross-language
  parity break versus the TS (number) and Rust (f64) reference. Changed
  Unit.ActionScope.Spend.maxSpend to Double and added a decimal-preserving
  parser used only for this currency field (asNum's other callers — token/byte
  counts — are genuinely integer-only and are unchanged).

+++ java/kcp-planner/src/main/java/no/cantara/kcp/planner/model/Unit.java
@@ -68,12 +68,16 @@ public record Unit(
         /**
          * Spend limits declared under a skill's action scope.
          *
-         * @param maxSpend       the maximum spend the skill may authorize
+         * @param maxSpend       the maximum spend the skill may authorize — a
+         *                       currency amount (e.g. 4.99), not a count, so this
+         *                       is a Double (mirrors the TS `number` and Rust
+         *                       `f64` reference/port; do not change to Long/Long,
+         *                       which would silently truncate fractional ceilings)
          * @param allowedVendors vendors the skill may spend with
          * @param currency       the settlement currency
          */
         public record Spend(
-                Long maxSpend,
+                Double maxSpend,
                 List<String> allowedVendors,
                 String currency) {
         }
